+++ iosApp/iosApp/ui/util/UnixTimeFormatter.swift
@@ -2,10 +2,10 @@ import Foundation
 
 class UnixTimeFormatter {
     func unixTimeToString(
-        unixTime: Int64,
+        unixTime: Double,
         format: String
     ) -> String {
-        let date = Date(timeIntervalSince1970: Double(unixTime))
+        let date = Date(timeIntervalSince1970: unixTime)
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = format
         return dateFormatter.string(from: date)
